requirements_production.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1
Flask-Session==0.5.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
openai==0.28.1
sendgrid==6.10.0
requests==2.31.0
Werkzeug==2.3.7
WTForms==3.0.1
bleach==6.0.0
psutil==5.9.5
redis==4.6.0